refactor: make the error parameter on subscribe non optional
