+++ packages/realtime_client/lib/realtime_client.dart
@@ -1,5 +1,6 @@
-export 'src/realtime_channel.dart' hide ToType;
+export 'src/constants.dart' show RealtimeLogLevel;
+export 'src/realtime_channel.dart';
 export 'src/realtime_client.dart';
 export 'src/realtime_presence.dart';
 export 'src/transformers.dart' hide getEnrichedPayload, getPayloadRecords;
-export 'src/constants.dart' show RealtimeLogLevel;
+export 'src/types.dart' hide ToType;
