+++ packages/realtime_client/lib/src/push.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
+import 'package:realtime_client/realtime_client.dart';
 import 'package:realtime_client/src/constants.dart';
 import 'package:realtime_client/src/message.dart';
-import 'package:realtime_client/src/realtime_channel.dart';
 
 typedef Callback = void Function(dynamic response);
