+++ packages/realtime_client/lib/src/realtime_channel.dart
@@ -7,104 +7,7 @@ import 'package:realtime_client/src/constants.dart';
 import 'package:realtime_client/src/push.dart';
 import 'package:realtime_client/src/retry_timer.dart';
 import 'package:realtime_client/src/transformers.dart';
-
-typedef BindingCallback = void Function(dynamic payload, [dynamic ref]);
-
-class Binding {
-  String type;
-  Map<String, String> filter;
-  BindingCallback callback;
-  String? id;
-
-  Binding(
-    this.type,
-    this.filter,
-    this.callback, [
-    this.id,
-  ]);
-
-  Binding copyWith({
-    String? type,
-    Map<String, String>? filter,
-    BindingCallback? callback,
-    String? id,
-  }) {
-    return Binding(
-      type ?? this.type,
-      filter ?? this.filter,
-      callback ?? this.callback,
-      id ?? this.id,
-    );
-  }
-}
-
-class ChannelFilter {
-  final String? event;
-  final String? schema;
-  final String? table;
-  final String? filter;
-
-  ChannelFilter({
-    this.event,
-    this.schema,
-    this.table,
-    this.filter,
-  });
-
-  Map<String, String> toMap() {
-    return {
-      if (event != null) 'event': event!,
-      if (schema != null) 'schema': schema!,
-      if (table != null) 'table': table!,
-      if (filter != null) 'filter': filter!,
-    };
-  }
-}
-
-enum ChannelResponse { ok, timedOut, rateLimited }
-
-enum RealtimeListenTypes { postgresChanges, broadcast, presence }
-
-extension ToType on RealtimeListenTypes {
-  String toType() {
-    if (this == RealtimeListenTypes.postgresChanges) {
-      return 'postgres_changes';
-    } else {
-      return name;
-    }
-  }
-}
-
-class RealtimeChannelConfig {
-  /// [ack] option instructs server to acknowlege that broadcast message was received
-  final bool ack;
-
-  /// [self] option enables client to receive message it broadcasted
-  final bool self;
-
-  /// [key] option is used to track presence payload across clients
-  final String key;
-
-  const RealtimeChannelConfig({
-    this.ack = false,
-    this.self = false,
-    this.key = '',
-  });
-
-  Map<String, dynamic> toMap() {
-    return {
-      'config': {
-        'broadcast': {
-          'ack': ack,
-          'self': self,
-        },
-        'presence': {
-          'key': key,
-        },
-      }
-    };
-  }
-}
+import 'package:realtime_client/src/types.dart';
 
 class RealtimeChannel {
   final Map<String, List<Binding>> _bindings = {};
@@ -189,7 +92,7 @@ class RealtimeChannel {
   ///
   /// [timeout] parameter can be used to override the default timeout set on `RealtimeClient`.
   void subscribe([
-    void Function(String status, [Object? error])? callback,
+    void Function(String status, Object? error)? callback,
     Duration? timeout,
   ]) {
     if (joinedOnce == true) {
@@ -202,7 +105,7 @@ class RealtimeChannel {
         if (callback != null) callback('CHANNEL_ERROR', e);
       });
       onClose(() {
-        if (callback != null) callback('CLOSED');
+        if (callback != null) callback('CLOSED', null);
       });
 
       final accessTokenPayload = <String, String>{};
@@ -229,7 +132,7 @@ class RealtimeChannel {
           if (socket.accessToken != null) socket.setAuth(socket.accessToken);
 
           if (serverPostgresFilters == null) {
-            if (callback != null) callback('SUBSCRIBED');
+            if (callback != null) callback('SUBSCRIBED', null);
             return;
           } else {
             final clientPostgresBindings = _bindings['postgres_changes'];
@@ -268,7 +171,7 @@ class RealtimeChannel {
 
             _bindings['postgres_changes'] = newPostgresBindings;
 
-            if (callback != null) callback('SUBSCRIBED');
+            if (callback != null) callback('SUBSCRIBED', null);
             return;
           }
         },
@@ -285,7 +188,7 @@ class RealtimeChannel {
         }
         return;
       }).receive('timeout', (_) {
-        if (callback != null) callback('TIMED_OUT');
+        if (callback != null) callback('TIMED_OUT', null);
         return;
       });
     }
